requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
supabase>=2.3.0
PyPDF2>=3.0.0
spire.doc>=8.6.0
requests>=2.31.0 